Add "test prod" stage, variable for IP address, and update Docker logout command
Created a new "test prod" stage that performs a curl request
to http://<ip-add-aws>, followed by an echo of the HTTP status code (200 or not).
This allows for a simple verification of the deployment.

Additionally, introduced a variable for the IP address to ensure it's
defined in only one place, making future updates easier. The Docker
logout command was also updated to be executed via SSH, improving consistency across stages.